+++ ui/src/app/core/auth/auth.guard.ts
@@ -2,6 +2,7 @@ import { AuthService } from '@/app/core/auth/auth.service'
 import { SettingsService } from '@/app/core/settings.service'
 import { Injectable } from '@angular/core'
 import { ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot } from '@angular/router'
+import { firstValueFrom } from 'rxjs'
 
 @Injectable({
   providedIn: 'root',
@@ -19,7 +20,7 @@ export class AuthGuard implements CanActivate {
   ): Promise<boolean> {
     // ensure app settings are loaded
     if (!this.$settings.settingsLoaded) {
-      await this.$settings.onSettingsLoaded.toPromise()
+      await firstValueFrom(this.$settings.onSettingsLoaded)
     }
 
     if (this.$auth.isLoggedIn()) {