+++ ui/src/app/modules/platform-tools/docker/startup-script/startup-script.resolver.ts
@@ -2,6 +2,7 @@ import { ApiService } from '@/app/core/api.service'
 import { Injectable } from '@angular/core'
 import { Resolve, Router } from '@angular/router'
 import { ToastrService } from 'ngx-toastr'
+import { firstValueFrom } from 'rxjs'
 
 @Injectable()
 export class StartupScriptResolver implements Resolve<any> {
@@ -13,7 +14,7 @@ export class StartupScriptResolver implements Resolve<any> {
 
   async resolve() {
     try {
-      return await this.$api.get('/platform-tools/docker/startup-script').toPromise()
+      return await firstValueFrom(this.$api.get('/platform-tools/docker/startup-script'))
     } catch (err) {
       this.$toastr.error(err.message, 'Failed to Load Startup Script')
       this.$router.navigate(['/'])