+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 - no longer use `webpack`
   - the final size of the `npm` package is actually higher with `webpack` than without
 - update dependencies, overrides, updated lint
+- `.toPromise()` is deprecated, change to `firstValueFrom()`
 
 ## v4.58.0 (2024-09-08)